refactor(response): Move all response handling to one module
