+++ src/nerve/github/entities.rs
@@ -1,5 +1,4 @@
 // Github entities represented as structs.
-use std::fmt;
 use rustc_serialize::json;
 
 /// An Issue-to-be. It doesn't have number, state or any timestamp because
@@ -18,7 +17,6 @@ pub struct NewIssue {
 }
 
 pub type NewIssues = Vec<NewIssue>;
-
 
 
 /// A partial representation of a Github Issue. The represented fields are the
@@ -56,40 +54,4 @@ pub struct Label {
 #[derive(Debug, RustcDecodable, RustcEncodable)]
 pub struct User {
     pub login: String,
-}
-
-
-#[derive(Debug, RustcDecodable, RustcEncodable)]
-pub struct Error {
-    pub message: String,
-    pub errors: Vec<ErrorResource>,
-}
-
-impl Error {
-    pub fn from_str(data: &str) -> Result<Error, json::DecoderError> {
-        json::decode(data)
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let error = self.errors.first().unwrap();
-
-        write!(f, "the field '{}' {}", error.field, "has an invalid value.")
-    }
-}
-
-#[derive(Debug, RustcDecodable, RustcEncodable)]
-pub struct ErrorResource {
-    pub code: String,
-    pub resource: String,
-    pub field: String,
-}
-
-#[derive(Debug, RustcDecodable, RustcEncodable)]
-pub enum ErrorName {
-    Invalid,
-    Missing,
-    MissingField,
-    AlreadyExists,
 }
